Move TCP dialchain builder into TCP package
